Add checks for the timestep type (fixed vs adjusted).
The private boolean `adjustableTimestep_` gets its value by reading the respective value on the `controlDict` dictionary. This is expected to work in all the OpenFOAM versions.

At the same time, a private copy of the runTime object is added. This is going to be useful also for other features.

The `README.md` is updated to drop the requirement for adjustable timestep.