Add the reading and checking of the adapter's YAML config file.
The adapter expects a YAML file with the name "precice-adapter-config.yml" and form:

```
participant: Fluid

precice-config-file: ../precice-config.xml

interfaces:
- mesh: Fluid-Mesh
  patches: [interface]
  write-data: Temperature
  read-data: Heat-Flux

subcycling: Yes
checkpointing: Yes
```

The adapter currently does not read the interfaces, as the respective class needs to get implemented first.

The nodes are checked one-by-one for existence, in order to allow for special checks for each, but this could be simplified if needed.

A new function `errorAndExit()` handles the errors, but currently stops the simulation without a failure exit code. The checking method does not exit on the first error, in order to find all the problems in the file and make the user's life easier.