Checkpointing works, teardown improvements.
Checkpointing:
looking for all the registered objects of type volScalarField, volVectorField, surfaceScalarField and adding them in the list of checkpointed fields.

A new method `setupCheckpointing()` finds and adds the fields.

In the `execute()`, checkpoint reading needed to come before reading the new coupling data, since it was overwriting them.

Teardown:
Trying to handle better the cases when the coupling finishes before the solver or the opposite. There is still a problem when the solver finishes while the coupling is ongoing.