Improvements in error handling.
Errors produced inside a functionObject's read() are caught as warnings and OpenFOAM does not exit. It simply ignores the functionObject.

Now the errors from `read()` are handled in `execute()`, which can normally through errors.

A separate `teardown()` method is introduced, which properly deletes the allocated memory. It is currently called only in the destructor.