Prevent early exits from the solver.
By default, when using implicit coupling, set the solver's endTime to infinity to only control it with preCICE.

The functionObjects' end() methods are not being called automatically in this way, so the adapter triggers them automatically at the end of the coupling.

The user can disable these automations in the configuration file, by setting `preventEarlyExit: No`.

Furthermore, finalization became smoother.