Tutorial: adjust solvers' endTime.
When the adapter is loaded, it will overwrite the entTime of the solver. However, if something goes wrong and the adapter is not loaded, the simulation will continue without the adapter. Reduced the endTime to avoid the waiting time in this case. Set it to be the same as the preCICE endTime.