Create each interface's data buffer according to the data type.
Each interface has one data buffer, shared between all the assigned coupling
data users. Up to now, this buffer was created in the interface's constructor
and it was hard-coded for scalar data. A comment was suggesting to resize
the buffer if a coupling data user contained vector data.

With this commit, the buffer is created by a different method,
which is called inside the Adapter.C, immediately after all the
coupling data readers/writers are added. It check if a reader/writer
has vector data and it creates a buffer of the respective size.

When writing scalar data, it works correctly, as only the first 1/3
of the buffer elements are used, both by the (current, CHT) coupling
data users and by preCICE. This has been tested with the included
tutorial case, but it has not yet been tested for any vector data.

Reminder: an interface is defined between two participants and it can
have multiple OpenFOAM boundary patches. Furthermore, multiple kind
of data (coupling data users) may be exchanged over an interface, which
can be of different types (scalar/vector).