Support nearest-projection mapping (#46)
Add support for nearest-projection mapping

Closes #2.

Commit messages of the squashed commits:

* Added the functionality of defining triangles for a 3D coupling mesh in order to use a nearest-projection mapping.
This includes the definition of two meshes for each participant: Face centers and face nodes. For the face nodes, data is interpolated from the face centers.
The definition of two meshes needs two different mesh names. Therefore, an example setup (with the respective naming) for a nearest-projection setup can be found in the tutorial case.
This version is currently a prototype.

* Adapted Formatting

* Add triSurface include path in Make/options

* Applied changes from PR review
Opened up a new tutorial case for nearest-projection mapping

* Restore original tutorial case

* Removed *vtk result files

* Bugfix in naming

* New setup/setting for nearest-projection mapping
Two meshes are defined in the solvers yaml files
additionally, a provideMeshConnectivity option needs to be set.
if you choose the provideMeshConnectivity option true, it is assumed,
that you want to provide the information on your write-data set
(consistent mapping). Therefore, the locationsType should be faceNodes.
Have a look in the tutorial case.

* Avoid error message due to vtk interface output option

* Provide Mesh Connectivity information and interpolate, if necessary

* Apply suggestions from code review

Co-Authored-By: DavidSCN <dav.schneider@tum.de>

* Added missing implementation/interpolation for these quantities
this enables support to write them with nearest-projection mapping

* Rearranged the interpolation step and extended content to fully support the CHT
module, cleanup and added several comments in the code. Associated the
meshConnectivity to each interface.

* Rearranged the tutorial case to be consistent with the existing tutorial
case. Still need to add a README.md

* Corrected setup for solid participant

* Added parallel version

* Case description in the README.md file

* Removed Info in the FSI cases and added comment

* Added error for wrong configuration in the solver files

* Remove stray blockMeshDict files

* Fix error handling in nearest-projection config

* Nearest-Projection README: Language edits

* Nearest Projection: Comments and formatting

* Fix provideMeshConnectivity existence check

* Nearest Projection tutorial formatting

* NP tutorial: Use implicit coupling for both serial and parallel

Note that the coupling scheme needed to be adjusted, so that the
heat flux is exchanged over the Solid-Mesh-Nodes, as this is used
now in a read-consistent mapping on the Fluid side.

* Added comment for `<export..` tag