Rephrase "unknown solver type" message
In normal behavior, the user should not need to manually specify the solver
type, but the adapter should determine it by itself. However, this means that
the user gets an "Unknown solver type" message by default.

This removes the "Unknown solver type" and adds a hint for possible
options to override it. This is raised both in the case when the solver
type is not specified, as well as in case the solver type is specified
but now known.