Replace non-const boundaryFieldRef wherever possible (#132)
* Refactor to reduce assigning fields at creation

This replaces statements such as:

    scalarField one_field = another_field;

by constructors:

    scalarField one_field(another_field);

This can be safer and faster, as it avoids implicit conversions in between.

* Replace non-const boundaryFieldRef wherever possible

OpenFOAM provides the methods boundaryField and boundaryFieldRef
to return references to the boundary of a field. Their difference
is that boundaryField returns a const reference, while boundaryFieldRef
returns a non-const reference.

There are a few parts of the code where we were using the non-const versions
where the const would be enough. This replaces boundaryFieldRef by
boundaryField, essentially keeping only one non-const instance at the read()
method of every coupling data reader.

Co-authored-by: David Schneider <dav.schneider@tum.de>